rename Sharing to SharingPage
